packages/api/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.41.0
python-multipart==0.0.26
presidio-analyzer==2.2.359
presidio-anonymizer==2.2.362
spacy==3.8.11
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
transformers==4.57.6
torch==2.10.0
sentence-transformers==5.2.3
optimum[openvino]==2.1.0
langchain==1.2.14
langchain-openai==1.1.14
langchain-anthropic==1.4.0
langchain-community==0.4.1
langchain-ollama==1.0.1
langchain-text-splitters==1.1.2
langchain-experimental==0.4.1
faiss-cpu==1.13.2
langdetect==1.0.9
lingua-language-detector==2.2.0
pymupdf==1.27.2
python-docx==1.2.0
openpyxl==3.1.5
odfpy==1.4.1
pandas==2.3.3
python-pptx==1.0.2
paddlepaddle==3.2.2
paddleocr==3.4.0
opencv-python==4.13.0.92
pillow==12.2.0
openai-whisper==20250625
ffmpeg-python==0.2.0
beautifulsoup4==4.13.5
lxml==6.0.2
markdown-it-py==4.0.0
pyyaml==6.0.2
extract-msg==0.55.0
ebooklib==0.20
striprtf==0.0.29
python-magic==0.4.27
pydantic-settings==2.13.1
cryptography==46.0.7
sqlalchemy==2.0.48
aiosqlite==0.22.1
asyncpg==0.30.0
alembic==1.15.2
redis[hiredis]==6.2.0
PyJWT==2.12.0
bcrypt==4.3.0
slowapi==0.1.9
email-validator==2.2.0
python-json-logger==3.3.0
prometheus-client==0.22.1
greenlet==3.3.2
pytest==9.0.3
pytest-asyncio==1.3.0
httpx==0.28.1
rank-bm25==0.2.2
pdfplumber==0.11.5
